+++ examples/e2e/requirements.txt
@@ -1,3 +1,4 @@
 Flask==1.1.1
 pytest==5.4.1
 requests==2.23.0
+testcontainers==3.3.0
