+++ examples/e2e/tests/conftest.py
@@ -1,4 +1,8 @@
 
+from testcontainers.compose import DockerCompose
+
+import pytest
+
 def pytest_addoption(parser):
     parser.addoption(
         "--publish-pact", type=str, action="store",
@@ -9,3 +13,25 @@ def pytest_addoption(parser):
         "--provider-url", type=str, action="store",
         help="The url to our provider."
     )
+
+
+@pytest.fixture(scope='session', autouse=True)
+def broker(request):
+    version = request.config.getoption('--publish-pact')
+    publish = True if version else False
+
+    # yield
+    if not publish:
+        yield
+        return
+
+    print('Starting broker')
+    with DockerCompose("../broker",
+                       compose_file_name=["docker-compose.yml"],
+                       pull=True) as compose:
+
+        stdout, stderr = compose.get_logs()
+        if stderr:
+            print("Errors\\n:{}".format(stderr))
+        print(stdout)
+        yield
