+++ examples/e2e/tests/consumer/userserviceclient-userservice.json
@@ -8,43 +8,6 @@
   "interactions": [
     {
       "description": "a request for UserA",
-      "providerState": "UserA exists and is not an administrator",
-      "request": {
-        "method": "get",
-        "path": "/users/UserA"
-      },
-      "response": {
-        "status": 200,
-        "headers": {
-        },
-        "body": {
-          "name": "UserA",
-          "id": "fc763eba-0905-41c5-a27f-3934ab26786c",
-          "created_on": "2016-12-15T20:16:01",
-          "ip_address": "127.0.0.1",
-          "admin": false
-        },
-        "matchingRules": {
-          "$.body": {
-            "match": "type"
-          },
-          "$.body.id": {
-            "match": "regex",
-            "regex": "[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}"
-          },
-          "$.body.created_on": {
-            "match": "regex",
-            "regex": "\\d+-\\d+-\\d+T\\d+:\\d+:\\d+"
-          },
-          "$.body.ip_address": {
-            "match": "regex",
-            "regex": "(\\d{1,3}\\.)+\\d{1,3}"
-          }
-        }
-      }
-    },
-    {
-      "description": "a request for UserA",
       "providerState": "UserA does not exist",
       "request": {
         "method": "get",
