+++ examples/e2e/tests/provider/test_provider.py
@@ -2,7 +2,6 @@
 
 import logging
 import os
-from testcontainers.compose import DockerCompose
 
 from pact import Verifier
 import pytest
@@ -21,31 +20,25 @@ PACT_BROKER_USERNAME = "pactbroker"
 PACT_BROKER_PASSWORD = "pactbroker"
 
 PACT_MOCK_HOST = 'localhost'
-PACT_MOCK_PORT = 1235
+PACT_MOCK_PORT = 5001
 PACT_URL = "http://{}:{}".format(PACT_MOCK_HOST, PACT_MOCK_PORT)
 PACT_DIR = os.path.dirname(os.path.realpath(__file__))
 
+@pytest.fixture
+def default_opts():
+    return {
+        'broker_username': PACT_BROKER_USERNAME,
+        'broker_password': PACT_BROKER_PASSWORD,
+        'broker_url': PACT_BROKER_URL
+    }
 
-def test_get_user_non_admin():
+
+def test_get_user_non_admin(default_opts):
     verifier = Verifier(provider='UserService',
                         provider_base_url=PACT_URL)
 
-    output, logs = verifier.verify_pacts('./userserviceclient-userservice.json',
-                                         verbose=False,
-                                         provider_states_setup_url="{}/_pact/provider_states".format(PACT_URL))
+    output, logs = verifier.verify_with_broker(**default_opts,
+                                               verbose=True,
+                                               provider_states_setup_url="{}/_pact/provider_states".format(PACT_URL))
 
     assert (output == 0)
-
-
-# @pytest.fixture(scope='session')
-# def broker():
-#     print('Starting broker')
-#     with DockerCompose("../broker",
-#                        compose_file_name=["docker-compose.yml"],
-#                        pull=True) as compose:
-
-#         stdout, stderr = compose.get_logs()
-#         if stderr:
-#             print("Errors\\n:{}".format(stderr))
-#         print(stdout)
-#         yield
